Fix package structure
