+++ src/main/java/org/example/logscanner/PipelineApplication.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.logscanner;
 
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.io.TextIO;
@@ -6,8 +6,8 @@ import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Flatten;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.example.transform.ExtractLogMetrics;
-import org.example.transform.ParseLog;
+import org.example.logscanner.transform.ExtractLogMetrics;
+import org.example.logscanner.transform.ParseLog;
 
 public class PipelineApplication {
