+++ src/main/java/org/example/logscanner/entity/LogElement.java
@@ -1,4 +1,4 @@
-package org.example.entity;
+package org.example.logscanner.entity;
 
 import lombok.Builder;
 import lombok.Data;
