+++ src/main/java/org/example/logscanner/transform/FormatKVMaps.java
@@ -1,4 +1,4 @@
-package org.example.transform;
+package org.example.logscanner.transform;
 
 import avro.shaded.com.google.common.collect.Lists;
 import org.apache.beam.sdk.transforms.SimpleFunction;
