+++ src/main/java/org/example/logscanner/transform/TopActiveIps.java
@@ -1,11 +1,11 @@
-package org.example.transform;
+package org.example.logscanner.transform;
 
 import org.apache.beam.sdk.transforms.*;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TypeDescriptor;
-import org.example.comparator.CompareByValue;
-import org.example.entity.LogElement;
-import org.example.util.FormatUtils;
+import org.example.logscanner.comparator.CompareByValue;
+import org.example.logscanner.entity.LogElement;
+import org.example.logscanner.util.FormatUtils;
 
 public class TopActiveIps extends PTransform<PCollection<LogElement>, PCollection<String>> {
