+++ src/test/java/org/example/logscanner/transform/FormatKVMapsTest.java
@@ -1,4 +1,4 @@
-package org.example.transform;
+package org.example.logscanner.transform;
 
 import org.apache.beam.sdk.testing.PAssert;
 import org.apache.beam.sdk.testing.TestPipeline;
@@ -9,23 +9,25 @@ import org.apache.beam.sdk.values.PCollection;
 import org.junit.Rule;
 import org.junit.Test;
 
-public class FormatKVTest {
+import java.util.List;
+
+public class FormatKVMapsTest {
 
     @Rule
     public final transient TestPipeline testPipeline = TestPipeline.create();
 
     @Test
-    public void givenValidKV_whenFormatting_thenOutputFormattedString() {
+    public void givenValidKVWithValueIterable_whenFormatting_thenOutputFormattedString() {
 
         // Mock
-        KV<String, String> inputKV = KV.of("testKey", "testValue");
+        KV<String, Iterable<String>> inputKV = KV.of("testKey", List.of("value1", "value2"));
 
-        String expectedOutput = "testKey: testValue";
+        String expectedOutput = "testKey: [value1, value2]";
 
         // Invoke
-        PCollection<KV<String, String>> input = testPipeline.apply("Create input", Create.of(inputKV));
+        PCollection<KV<String, Iterable<String>>> input = testPipeline.apply("Create input", Create.of(inputKV));
 
-        PCollection<String> output = input.apply("Format KV", MapElements.via(new FormatKV()));
+        PCollection<String> output = input.apply("Format KV with iterable", MapElements.via(new FormatKVMaps()));
 
         // Assert
         PAssert.that(output).containsInAnyOrder(expectedOutput);
@@ -33,4 +35,5 @@ public class FormatKVTest {
         testPipeline.run().waitUntilFinish();
 
     }
+
 }
